% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigmaAR.R
\name{SigmaAR}
\alias{SigmaAR}
\title{Simulate Covariance Matrix with an Auto-regressive (AR) Model}
\usage{
SigmaAR(p, phi)
}
\arguments{
\item{p}{Number of variables (rows x columns).}

\item{phi}{Auto-regression coefficient.}
}
\value{
A list with the values:
\item{\code{Sigma}}{A covariance matrix.}
}
\description{
Helper function to simulate synthetic data
}
\author{
Prof. Juan G. Colonna, PhD. \email{juancolonna@icomp.ufam.edu.br}

Prof. Marcelo Ruiz, PhD. \email{mruiz@exa.unrc.edu.ar}
}
