% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigmaAR.R
\name{SigmaAR}
\alias{SigmaAR}
\title{Simulate Covariance Matrix with an Auto-regressive (AR) Model}
\usage{
SigmaAR(p, phi)
}
\arguments{
\item{p}{Number of variables (rows x columns).}

\item{phi}{Auto-regression coefficient.}
}
\value{
A list with the values:
\item{\code{Sigma}}{A covariance matrix.}
}
\description{
Helper function to simulate Simulate Gaussian Data with an Autoregressive (AR) Model
}
\examples{
library(FastStepGraph)
library(MASS) # mvrnorm
phi = 0.4
p = 100  # Dimension
n = 100 # Sample size
Sigma = SigmaAR(p, phi) # Simulate Gaussian Data with Autoregressive (AR) Model
Omega = solve(Sigma)
Omega[abs(Omega) < 1e-5] = 0
X = list() # Generate Data from a Gaussian distribution
X = MASS::mvrnorm(n, mu=rep(0,p), Sigma)
X = scale(X)

}
\author{
Prof. Juan G. Colonna, PhD. \email{juancolonna@icomp.ufam.edu.br}

Prof. Marcelo Ruiz, PhD. \email{mruiz@exa.unrc.edu.ar}
}
